♻️ refactor: separate showAnimal function to other file
