+++ scripts/app.mjs
@@ -1,24 +1,11 @@
 import { ANIMALS } from './animals.mjs';
-
-const animalImage = document.getElementById('img');
-const animalName = document.getElementById('animal-name');
-const animalType = document.getElementById('animal-type');
-const animalDescription = document.getElementById('description');
+import { showAnimal } from './showAnimal.mjs';
 
 const previousButton = document.getElementById('prev-btn');
 const nextButton = document.getElementById('next-btn');
 const surpriseButton = document.getElementById('surprise-btn');
 
 let currentAnimal = 0;
-
-const showAnimal = (current) => {
-  const animal = ANIMALS[current];
-
-  animalImage.src = animal.image;
-  animalName.textContent = animal.name;
-  animalType.textContent = animal.type;
-  animalDescription.textContent = animal.text;
-};
 
 previousButton.addEventListener('click', () => {
   currentAnimal -= 1;
